requirements.txt
Flask==2.3.2
oracledb==1.4.2
python-dotenv==1.0.0